restrict view of attached MFs with the same rule as for template MFs
